+++ DayPicker.js
@@ -1,11 +1,10 @@
+/* eslint-disable no-var */
 /* eslint-env node */
 
-"use strict";
-
-var DayPicker = require("./lib/DayPicker");
-var DateUtils =  require("./lib/DateUtils");
-var LocaleUtils = require("./lib/LocaleUtils");
+var DayPicker = require('./lib/DayPicker');
+var DateUtils = require('./lib/DateUtils');
+var LocaleUtils = require('./lib/LocaleUtils');
 
 module.exports = DayPicker.default || DayPicker;
 module.exports.DateUtils = DateUtils.default || DateUtils;
-module.exports.LocaleUtils  = LocaleUtils.default || LocaleUtils;
+module.exports.LocaleUtils = LocaleUtils.default || LocaleUtils;