+++ src/DayPicker.js
@@ -1,46 +1,28 @@
-import React, { Component, PropTypes } from "react";
-import * as Helpers from "./Helpers";
-import * as DateUtils from "./DateUtils";
-import * as LocaleUtils from "./LocaleUtils";
+import React, { Component, PropTypes } from 'react';
 
-const keys = {
-  LEFT: 37,
-  UP: 38,
-  RIGHT: 39,
-  DOWN: 40,
-  ENTER: 13,
-  SPACE: 32
-};
+import Caption from './Caption';
+import Navbar from './Navbar';
+import Month from './Month';
+import Day from './Day';
 
-class Caption extends Component { // eslint-disable-line
+import * as Helpers from './Helpers';
+import * as DateUtils from './DateUtils';
+import * as LocaleUtils from './LocaleUtils';
 
-  render() {
-    const { date, locale, localeUtils, onClick } = this.props;
-    return (
-      <div className="DayPicker-Caption" onClick={ onClick }>
-        { localeUtils.formatMonthTitle(date, locale) }
-      </div>
-    );
-  }
-}
+import keys from './keys';
+import DayPickerPropTypes from './PropTypes';
 
 export default class DayPicker extends Component {
-  static VERSION = "2.0.0-beta";
+  static VERSION = '2.0.0-beta';
 
   static propTypes = {
-    tabIndex: PropTypes.number,
     initialMonth: PropTypes.instanceOf(Date),
     numberOfMonths: PropTypes.number,
 
     modifiers: PropTypes.object,
 
     locale: PropTypes.string,
-    localeUtils: PropTypes.shape({
-      formatMonthTitle: PropTypes.func,
-      formatWeekdayShort: PropTypes.func,
-      formatWeekdayLong: PropTypes.func,
-      getFirstDayOfWeek: PropTypes.func
-    }),
+    localeUtils: DayPickerPropTypes.localeUtils,
 
     enableOutsideDays: PropTypes.bool,
     canChangeMonth: PropTypes.bool,
@@ -48,47 +30,68 @@ export default class DayPicker extends Component {
     fromMonth: PropTypes.instanceOf(Date),
     toMonth: PropTypes.instanceOf(Date),
 
+    onKeyDown: PropTypes.func,
     onDayClick: PropTypes.func,
+    onDayKeyDown: PropTypes.func,
     onDayMouseEnter: PropTypes.func,
     onDayMouseLeave: PropTypes.func,
+    onDayTouchStart: PropTypes.func,
+    onDayTouchEnd: PropTypes.func,
     onMonthChange: PropTypes.func,
     onCaptionClick: PropTypes.func,
 
     renderDay: PropTypes.func,
 
-    captionElement: PropTypes.element
+    captionElement: PropTypes.element,
+
+    dir: PropTypes.string,
+    tabIndex: PropTypes.number,
+
   };
 
   static defaultProps = {
     tabIndex: 0,
     initialMonth: new Date(),
     numberOfMonths: 1,
-    locale: "en",
+    locale: 'en',
     localeUtils: LocaleUtils,
     enableOutsideDays: false,
     canChangeMonth: true,
     reverseMonths: false,
     renderDay: day => day.getDate(),
-    captionElement: <Caption />
+    captionElement: <Caption />,
   };
 
   constructor(props) {
     super(props);
-    this.state = {
-      currentMonth: Helpers.startOfMonth(props.initialMonth)
-    };
+
+    this.renderDayInMonth = this.renderDayInMonth.bind(this);
+    this.showNextMonth = this.showNextMonth.bind(this);
+    this.showPreviousMonth = this.showPreviousMonth.bind(this);
+
+    this.handleKeyDown = this.handleKeyDown.bind(this);
+    this.handleDayClick = this.handleDayClick.bind(this);
+    this.handleDayKeyDown = this.handleDayKeyDown.bind(this);
+
+    this.state = this.getStateFromProps(props);
   }
 
   componentWillReceiveProps(nextProps) {
     if (this.props.initialMonth !== nextProps.initialMonth) {
-      this.setState({
-        currentMonth: Helpers.startOfMonth(nextProps.initialMonth)
-      });
+      this.setState(this.getStateFromProps(nextProps));
     }
   }
 
+  getStateFromProps = props => ({
+    currentMonth: Helpers.startOfMonth(props.initialMonth),
+  })
+
+  getDayNodes() {
+    return this.refs.dayPicker.querySelectorAll('.DayPicker-Day:not(.DayPicker-Day--outside)');
+  }
+
   allowPreviousMonth() {
-    const previousMonth = DateUtils.addMonths(this.state.currentMonth, -1)
+    const previousMonth = DateUtils.addMonths(this.state.currentMonth, -1);
     return this.allowMonth(previousMonth);
   }
 
@@ -110,10 +113,7 @@ export default class DayPicker extends Component {
     if (!this.allowMonth(d)) {
       return;
     }
-
-    this.setState({
-      currentMonth: Helpers.startOfMonth(d)
-    }, callback);
+    this.setState({ currentMonth: Helpers.startOfMonth(d) }, callback);
   }
 
   showMonthAndCallHandler(d, callback) {
@@ -151,10 +151,6 @@ export default class DayPicker extends Component {
     this.showMonthAndCallHandler(nextMonth, callback);
   }
 
-  getDayNodes() {
-    return this.refs.dayPicker.querySelectorAll(".DayPicker-Day:not(.DayPicker-Day--outside)");
-  }
-
   focusFirstDayOfMonth() {
     this.getDayNodes()[0].focus();
   }
@@ -169,7 +165,7 @@ export default class DayPicker extends Component {
     const dayNodeIndex = [...dayNodes].indexOf(dayNode);
 
     if (dayNodeIndex === 0) {
-      this.showPreviousMonth(() => { this.focusLastDayOfMonth() })
+      this.showPreviousMonth(() => this.focusLastDayOfMonth());
     } else {
       dayNodes[dayNodeIndex - 1].focus();
     }
@@ -180,7 +176,7 @@ export default class DayPicker extends Component {
     const dayNodeIndex = [...dayNodes].indexOf(dayNode);
 
     if (dayNodeIndex === dayNodes.length - 1) {
-      this.showNextMonth(() => { this.focusFirstDayOfMonth() });
+      this.showNextMonth(() => this.focusFirstDayOfMonth());
     } else {
       dayNodes[dayNodeIndex + 1].focus();
     }
@@ -232,22 +228,22 @@ export default class DayPicker extends Component {
 
     if (canChangeMonth) {
       switch (e.keyCode) {
-      case keys.LEFT:
-        this.showPreviousMonth(onKeyDown);
-        break;
-      case keys.RIGHT:
-        this.showNextMonth(onKeyDown);
-        break;
-      case keys.UP:
-        this.showPreviousYear(onKeyDown);
-        break;
-      case keys.DOWN:
-        this.showNextYear(onKeyDown);
-        break;
-      default:
-        if (onKeyDown) {
-          onKeyDown(e);
-        }
+        case keys.LEFT:
+          this.showPreviousMonth(onKeyDown);
+          break;
+        case keys.RIGHT:
+          this.showNextMonth(onKeyDown);
+          break;
+        case keys.UP:
+          this.showPreviousYear(onKeyDown);
+          break;
+        case keys.DOWN:
+          this.showNextYear(onKeyDown);
+          break;
+        default:
+          if (onKeyDown) {
+            onKeyDown(e);
+          }
       }
     }
   }
@@ -255,57 +251,46 @@ export default class DayPicker extends Component {
   handleDayKeyDown(e, day, modifiers) {
     e.persist();
     switch (e.keyCode) {
-    case keys.LEFT:
-      Helpers.cancelEvent(e);
-      this.focusPreviousDay(e.target);
-      break;
-    case keys.RIGHT:
-      Helpers.cancelEvent(e);
-      this.focusNextDay(e.target);
-      break;
-    case keys.UP:
-      Helpers.cancelEvent(e);
-      this.focusPreviousWeek(e.target);
-      break;
-    case keys.DOWN:
-      Helpers.cancelEvent(e);
-      this.focusNextWeek(e.target);
-      break;
-    case keys.ENTER:
-    case keys.SPACE:
-      Helpers.cancelEvent(e);
-      if (this.props.onDayClick) {
-        this.handleDayClick(e, day, modifiers);
-      }
-      break;
+      case keys.LEFT:
+        Helpers.cancelEvent(e);
+        this.focusPreviousDay(e.target);
+        break;
+      case keys.RIGHT:
+        Helpers.cancelEvent(e);
+        this.focusNextDay(e.target);
+        break;
+      case keys.UP:
+        Helpers.cancelEvent(e);
+        this.focusPreviousWeek(e.target);
+        break;
+      case keys.DOWN:
+        Helpers.cancelEvent(e);
+        this.focusNextWeek(e.target);
+        break;
+      case keys.ENTER:
+      case keys.SPACE:
+        Helpers.cancelEvent(e);
+        if (this.props.onDayClick) {
+          this.handleDayClick(e, day, modifiers);
+        }
+        break;
+      default:
+        break;
     }
-  }
-
-  handleCaptionClick(e, currentMonth) {
-    e.persist();
-    this.props.onCaptionClick(e, currentMonth);
+    if (this.props.onDayKeyDown) {
+      this.props.onDayKeyDown(e, day, modifiers);
+    }
   }
 
   handleDayClick(e, day, modifiers) {
     e.persist();
-    if (modifiers.indexOf("outside") > -1) {
-      this.handleOutsideDayPress(day);
+    if (modifiers.indexOf('outside') > -1) {
+      this.handleOutsideDayClick(day);
     }
-
     this.props.onDayClick(e, day, modifiers);
   }
 
-  handleDayMouseEnter(e, day, modifiers) {
-    e.persist();
-    this.props.onDayMouseEnter(e, day, modifiers);
-  }
-
-  handleDayMouseLeave(e, day, modifiers) {
-    e.persist();
-    this.props.onDayMouseLeave(e, day, modifiers);
-  }
-
-  handleOutsideDayPress(day) {
+  handleOutsideDayClick(day) {
     const { currentMonth } = this.state;
     const { numberOfMonths } = this.props;
     const diffInMonths = Helpers.getMonthsDiff(currentMonth, day);
@@ -316,118 +301,23 @@ export default class DayPicker extends Component {
     }
   }
 
-  renderNavBar() {
-    const baseClass = "DayPicker-NavButton DayPicker-NavButton";
-    const isRTL = this.props.dir === "rtl";
-
-    const leftButton = isRTL ? this.allowNextMonth() : this.allowPreviousMonth();
-    const rightButton = isRTL ? this.allowPreviousMonth() : this.allowNextMonth();
-
-    return (
-      <div className="DayPicker-NavBar">
-        { leftButton &&
-          <span
-            key="left"
-            className={ `${baseClass}--prev` }
-            onClick={ () => isRTL ? this.showNextMonth() : this.showPreviousMonth() }
-          />
-        }
-        { rightButton &&
-          <span
-            key="right"
-            className={ `${baseClass}--next` }
-            onClick={ () => isRTL ? this.showPreviousMonth() : this.showNextMonth() }
-          />
-        }
-      </div>
-    );
-  }
-
-  renderMonth(date, i) {
-    const { locale, localeUtils, onCaptionClick, captionElement } = this.props;
-
-    const caption = React.cloneElement(captionElement, {
-      date, localeUtils, locale,
-      onClick: onCaptionClick ? e => this.handleCaptionClick(e, date) : null
-    });
-
-    return (
-      <div
-        className="DayPicker-Month"
-        key={ i }>
-
-        { caption }
-
-        <div className="DayPicker-Weekdays" role="rowgroup">
-          <div className="DayPicker-WeekdaysRow" role="columnheader">
-            { this.renderWeekDays() }
-          </div>
-        </div>
-        <div className="DayPicker-Body" role="rowgroup">
-          { this.renderWeeksInMonth(date) }
-        </div>
-      </div>
-    );
-  }
-
-  renderWeekDays() {
-    const { locale, localeUtils } = this.props;
-    const days = [];
-    for (let i = 0; i < 7; i++) {
-      days.push(
-        <div key={ i } className="DayPicker-Weekday">
-          <abbr title={ localeUtils.formatWeekdayLong(i, locale) }>
-            { localeUtils.formatWeekdayShort(i, locale) }
-          </abbr>
-        </div>
-      );
+  renderDayInMonth(day, month) {
+    let dayModifiers = [];
+    if (DateUtils.isSameDay(day, new Date())) {
+      dayModifiers.push('today');
     }
-    return days;
-  }
-
-  renderWeeksInMonth(month) {
-    const { locale, localeUtils } = this.props;
-    const firstDayOfWeek = localeUtils.getFirstDayOfWeek(locale);
-    return Helpers.getWeekArray(month, firstDayOfWeek).map((week, i) =>
-      <div key={ i } className="DayPicker-Week" role="row">
-        { week.map(day => this.renderDay(month, day)) }
-      </div>
-    );
-  }
-
-  renderDay(month, day) {
-
-    const { enableOutsideDays, modifiers: modifierFunctions } = this.props;
-
-    let className = "DayPicker-Day";
-    let modifiers = [];
-    const key = `${day.getFullYear()}${day.getMonth()}${day.getDate()}`;
-
-    const isToday = DateUtils.isSameDay(day, new Date());
-    if (isToday) {
-      modifiers.push("today");
+    if (day.getMonth() !== month.getMonth()) {
+      dayModifiers.push('outside');
+    }
+    if (this.props.modifiers) {
+      dayModifiers = dayModifiers.concat(
+        Helpers.getModifiersForDay(day, this.props.modifiers)
+      );
     }
 
     const isOutside = day.getMonth() !== month.getMonth();
-    if (isOutside) {
-      modifiers.push("outside");
-    }
-
-    if (modifierFunctions) {
-      const customModifiers = Helpers.getModifiersForDay(day, modifierFunctions);
-      modifiers = [...modifiers, ...customModifiers];
-    }
-
-    className += modifiers.map(modifier => ` ${className}--${modifier}`).join("");
-
-    if (isOutside && !enableOutsideDays) {
-      return <div key={ `outside-${key}` } className={ className } />;
-    }
-
-    const { onDayMouseEnter, onDayMouseLeave, onDayClick }
-      = this.props;
     let tabIndex = null;
-    if (onDayClick && !isOutside) {
+    if (this.props.onDayClick && !isOutside) {
       tabIndex = -1;
       // Focus on the first day of the month
       if (day.getDate() === 1) {
@@ -435,66 +325,93 @@ export default class DayPicker extends Component {
       }
     }
 
-    const { localeUtils, locale } = this.props;
-    const ariaLabel = localeUtils.formatDay(day, locale);
-    const ariaDisabled = isOutside ? "true" : "false";
-
+    const key = `${day.getFullYear()}${day.getMonth()}${day.getDate()}`;
     return (
-      <div key={ key } className={ className }
-        tabIndex={ tabIndex }
-        role="gridcell"
-        aria-label={ ariaLabel }
-        aria-disabled={ ariaDisabled }
-        onKeyDown={
-          (e) => this.handleDayKeyDown(e, day, modifiers) }
-        onMouseEnter= { onDayMouseEnter ?
-          (e) => this.handleDayMouseEnter(e, day, modifiers) : null }
-        onMouseLeave= { onDayMouseLeave ?
-          (e) => this.handleDayMouseLeave(e, day, modifiers) : null }
-        onClick= { onDayClick ?
-          (e) => this.handleDayClick(e, day, modifiers) : null }
-        >
-        { this.props.renderDay(day) }
-      </div>
+      <Day
+        key={`${isOutside ? 'outside-' : ''}${key}`}
+        day={day}
+        modifiers={dayModifiers}
+        empty={isOutside && !this.props.enableOutsideDays}
+        tabIndex={tabIndex}
+        ariaLabel={this.props.localeUtils.formatDay(day, this.props.locale)}
+        ariaDisabled={isOutside}
+        onMouseEnter={this.props.onDayMouseEnter}
+        onMouseLeave={this.props.onDayMouseLeave}
+        onKeyDown={this.handleDayKeyDown}
+        onTouchStart={this.props.onDayTouchStart}
+        onTouchEnd={this.props.onDayTouchEnd}
+        onClick={this.props.onDayClick ? this.handleDayClick : undefined}
+      >
+        {this.props.renderDay(day)}
+      </Day>
     );
   }
 
+  renderMonths() {
+    const months = [];
+    const firstDayOfWeek = this.props.localeUtils.getFirstDayOfWeek(this.props.locale);
+
+    for (let i = 0; i < this.props.numberOfMonths; i++) {
+      const month = DateUtils.addMonths(this.state.currentMonth, i);
+
+      months.push(
+        <Month
+          className="DayPicker-Month"
+          wrapperClassName="DayPicker-Body"
+          weekClassName="DayPicker-Week"
+          locale={this.props.locale}
+          localeUtils={this.props.localeUtils}
+          key={i}
+          month={month}
+          firstDayOfWeek={firstDayOfWeek}
+          captionElement={this.props.captionElement}
+          onCaptionClick={this.props.onCaptionClick}
+        >
+          {this.renderDayInMonth}
+        </Month>);
+    }
+
+    if (this.props.reverseMonths) {
+      months.reverse();
+    }
+    return months;
+  }
+
   render() {
-    const { numberOfMonths, locale, canChangeMonth, reverseMonths, ...attributes } = this.props;
-    const { currentMonth } = this.state;
+    const {
+      locale,
+      canChangeMonth,
+      onDayClick,
+    ...attributes } = this.props;
     let className = `DayPicker DayPicker--${locale}`;
 
-    if (!this.props.onDayClick) {
+    if (!onDayClick) {
       className = `${className} DayPicker--interactionDisabled`;
     }
     if (attributes.className) {
       className = `${className} ${attributes.className}`;
     }
 
-    const months = [];
-    let month;
-    for (let i = 0; i < numberOfMonths; i++) {
-      month = DateUtils.addMonths(currentMonth, i);
-      months.push(this.renderMonth(month, i));
-    }
-
-    if (reverseMonths) {
-      months.reverse();
-    }
-
     return (
       <div
         {...attributes}
-        className={ className }
+        className={className}
         ref="dayPicker"
-        role="widget"
-        tabIndex={ canChangeMonth && attributes.tabIndex }
-        onKeyDown={ e => this.handleKeyDown(e) }>
-        { canChangeMonth && this.renderNavBar() }
-        { months }
+        role="application"
+        tabIndex={canChangeMonth && attributes.tabIndex}
+        onKeyDown={this.handleKeyDown}
+      >
+        {canChangeMonth &&
+          <Navbar
+            showPreviousButton={this.allowPreviousMonth()}
+            showNextButton={this.allowNextMonth()}
+            onNextClick={this.showNextMonth}
+            onPreviousClick={this.showPreviousMonth}
+            dir={attributes.dir}
+          />
+        }
+        {this.renderMonths()}
       </div>
     );
   }
-
-
 }