+++ src/LocaleUtils.js
@@ -1,11 +1,11 @@
-const WEEKDAYS_LONG = ["Sunday", "Monday", "Tuesday",
-  "Wednesday", "Thursday", "Friday", "Saturday"];
+const WEEKDAYS_LONG = ['Sunday', 'Monday', 'Tuesday',
+  'Wednesday', 'Thursday', 'Friday', 'Saturday'];
 
-const WEEKDAYS_SHORT = ["Su", "Mo", "Tu",
-  "We", "Th", "Fr", "Sa"];
+const WEEKDAYS_SHORT = ['Su', 'Mo', 'Tu',
+  'We', 'Th', 'Fr', 'Sa'];
 
-const MONTHS = ["January", "February", "March", "April", "May", "June",
-  "July", "August", "September", "October", "November", "December"];
+const MONTHS = ['January', 'February', 'March', 'April', 'May', 'June',
+  'July', 'August', 'September', 'October', 'November', 'December'];
 
 export function formatDay(day) {
   return day.toDateString();
@@ -37,5 +37,5 @@ export default {
   formatWeekdayShort,
   formatWeekdayLong,
   getFirstDayOfWeek,
-  getMonths
-}
+  getMonths,
+};