+++ src/addons/LocaleUtils.js
@@ -1,2 +1,2 @@
 // Deprecated: remove from the next major version
-export default from "./MomentLocaleUtils";
+export default from './MomentLocaleUtils';